requirements.txt
Django==2.1
djangorestframework==3.9.2
PyJWT==1.7.1
PyMySQL==0.9.3
pytz==2019.1
SQLAlchemy==1.3.2
xlrd==1.2.0
xlwt==1.3.0
